[refactor]! Migrasi manajemen harga dari Variety ke SeedLot
- Hapus kolom 'price' dari tabel varieties (breaking change)
- Pindahkan pengelolaan harga sepenuhnya ke level SeedLot via 'price_per_unit'
- Ubah tipe data price_per_unit dari decimal(12,2) ke unsignedBigInteger untuk Rupiah
- Tambahkan accessor price_range dan has_price di model Variety untuk kalkulasi dinamis
- Hapus field price dari form create/edit Variety dan validasi controller
- Update API response untuk tidak lagi mengembalikan price_cents/price_idr dari Variety
- Sesuaikan OrderItem:: createFromVariety() agar SeedLot menjadi parameter wajib
- Tampilkan Price Range di tabel daftar varieties sebagai pengganti harga statis

+++ app/Models/SeedLot.php
@@ -25,7 +25,7 @@ class SeedLot extends Model
 
     protected $casts = [
         'quantity' => 'integer',
-        'price_per_unit' => 'float',
+        'price_per_unit' => 'integer',
         'is_sellable' => 'boolean',
         'production_year' => 'integer',
         'harvest_date' => 'date',
